requirements.txt
requests
discord
python-dotenv
python-dateutil